+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Skips parsing again if body already used (`res.bodyUsed`).  
   This also means `body` _won't be_ set `undefined` in cases where the
   Content-Type is not recognized.
+- Allows passing in an array of media types to be read and possibly parsed.  
+  Only JSON is actually parsed. The rest are read to `res.body`.
+- Reads `text/javascript` as regular text, not JSON.
 
 ## 0.1.338 (Dec 11, 2015)
 - Skips parsing when response 304 Not Modified.