+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 - Allows passing in an array of media types to be read and possibly parsed.  
   Only JSON is actually parsed. The rest are read to `res.body`.
 - Reads `text/javascript` as regular text, not JSON.
-- Prevents JSON parse errors if `Content-Length` is `0` by not parsing then.
+- Skips parsing JSON if no actual content is returned (such as in response to
+  a `HEAD`) and sets `body` to `undefined`.
+- Sets the `body` of a response (assigned to `err.response`) to the invalid JSON
+  should an error occur.
 
 ## 0.1.338 (Dec 11, 2015)
 - Skips parsing when response 304 Not Modified.